Change method for creating executable object
The most recent release of glibc no longer allow the creation of
executable shared libraries with the original approach (see
<https://stackoverflow.com/questions/59074126/loading-executable-or-executing-a-library>
for details). To work around this, the executable loader is explicitly
defined along with a library entry point function. This means that the
number of arguments and the argument array are not accessible, so to
work around this, the arguments are extracted from "/proc/self/cmdline".
To figure out the path of the linker, the build system compiles an empty
program then uses ldd(1) to find the Linux linker. I suspect this code
may have some portability issues due to the aforementioned kludges and
the fact that calling fprintf(3) — now replaced with dprintf(3) — caused
the program to crash when compiled on Ubuntu 20.04, so the executable
shared library may go away entirely in future.

This should resolve <https://github.com/ericpruitt/homeishome/issues/1>.

The address and undefined behavior sanitizers no longer work with the
program, so they have been removed for now.