Refactor build to separate executable and library
This modifies the build system so there are two ways the project can be
built: either as an executable or as a plain, non-executable shared
library. This should allow the project to be usable on systems where the
executable cannot be built or cannot be run successfully for whatever
reasons.