Use POSIX exit statuses when execvp(3) fails
Section 2.8.2 of POSIX issue 7, 2018 edition reads "If a command is not
found, the exit status shall be 127. If the command name is found, but
it is not an executable utility, the exit status shall be 126.
Applications that invoke utilities without using the shell should use
these exit status values to report similar errors." This change modifies
the executable binary so it acts accordingly.